change directory name
change directory name from "fullsnak" to "fullsnake"
