--- src/main/java/com/fullsnack/eimsfuhcmbe/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.fullsnack.eimsfuhcmbe.controller;
-
-public class UserController {
-}
