--- src/main/java/com/fullsnack/eimsfuhcmbe/entity/User.java
@@ -1,4 +1,0 @@
-package com.fullsnack.eimsfuhcmbe.entity;
-
-public class User {
-}
