--- src/main/java/com/fullsnack/eimsfuhcmbe/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package com.fullsnack.eimsfuhcmbe.repository;
-
-public class UserRepository {
-}
