--- src/main/java/com/fullsnack/eimsfuhcmbe/service/UserServiceImpl.java
@@ -1,4 +1,0 @@
-package com.fullsnack.eimsfuhcmbe.service;
-
-public class UserServiceImpl {
-}
