+++ src/main/java/com/fullsnacke/eimsfuhcmbe/EimsFuhcmBeApplication.java
@@ -1,4 +1,4 @@
-package com.fullsnack.eimsfuhcmbe;
+package com.fullsnacke.eimsfuhcmbe;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
