+++ src/main/java/com/fullsnacke/eimsfuhcmbe/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.fullsnacke.eimsfuhcmbe.controller;
+
+public class UserController {
+}
