+++ src/main/java/com/fullsnacke/eimsfuhcmbe/entity/User.java
@@ -1,0 +1,4 @@
+package com.fullsnacke.eimsfuhcmbe.entity;
+
+public class User {
+}
