+++ src/main/java/com/fullsnacke/eimsfuhcmbe/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.fullsnacke.eimsfuhcmbe.repository;
+
+public class UserRepository {
+}
