+++ src/main/java/com/fullsnacke/eimsfuhcmbe/service/UserService.java
@@ -1,0 +1,4 @@
+package com.fullsnacke.eimsfuhcmbe.service;
+
+public class UserService {
+}
