+++ src/main/java/com/fullsnacke/eimsfuhcmbe/service/UserServiceImpl.java
@@ -1,0 +1,4 @@
+package com.fullsnacke.eimsfuhcmbe.service;
+
+public class UserServiceImpl {
+}
